+++ src/util/raster.hh
@@ -22,8 +22,8 @@ TwoDStorage<uint8_t>::TwoDStorage( const unsigned int width, const unsigned int 
 class BaseRaster
 {
 protected:
-  unsigned int display_width_, display_height_;
-  unsigned int width_, height_;
+  uint16_t display_width_, display_height_;
+  uint16_t width_, height_;
 
   TwoD< uint8_t > Y_ { width_, height_ },
     U_ { width_ / 2, height_ / 2 },
@@ -32,8 +32,8 @@ protected:
   size_t raw_hash( void ) const;
 
 public:
-  BaseRaster( const unsigned int display_width, const unsigned int display_height,
-    const unsigned int width, const unsigned int height );
+  BaseRaster( const uint16_t display_width, const uint16_t display_height,
+    const uint16_t width, const uint16_t height );
 
   TwoD< uint8_t > & Y( void ) { return Y_; }
   TwoD< uint8_t > & U( void ) { return U_; }
@@ -43,13 +43,13 @@ public:
   const TwoD< uint8_t > & U( void ) const { return U_; }
   const TwoD< uint8_t > & V( void ) const { return V_; }
 
-  unsigned int width( void ) const { return width_; }
-  unsigned int height( void ) const { return height_; }
-  unsigned int display_width( void ) const { return display_width_; }
-  unsigned int display_height( void ) const { return display_height_; }
+  uint16_t width( void ) const { return width_; }
+  uint16_t height( void ) const { return height_; }
+  uint16_t display_width( void ) const { return display_width_; }
+  uint16_t display_height( void ) const { return display_height_; }
 
-  unsigned int chroma_display_width() const { return (1 + display_width_) / 2; }
-  unsigned int chroma_display_height() const { return (1 + display_height_) / 2; }
+  uint16_t chroma_display_width() const { return (1 + display_width_) / 2; }
+  uint16_t chroma_display_height() const { return (1 + display_height_) / 2; }
 
   // SSIM as determined by libx264
   double quality( const BaseRaster & other ) const;
